count codepoints instead of bytes, to determine width
A complete solution would be to count grapheme clusters, but that would
require adding a dependency on something like zg.

Counting codepoints will ensure that typical non-ASCII text is
supported, but you can still throw it off with more complex Unicode
constructions, which might not be so useful in help text.

Fixes #75